v1.3 (#6)
* Set fullscreen

* Changes in buttons layout.

* Updates for gradle 9.4.1 and kotlin 2.4.
Change com.github.alpbak:BoxedVerticalSeekBar:1.1.1 for
com.github.ML-SPD:BoxedVerticalSeekBar:v1.1.1-fix5 due to some
compilation errors.
Remove some useless code.

+++ app/src/main/java/org/gnu/itsmoroto/midandpad/CCBar.kt
@@ -96,7 +96,7 @@ class CCBar : BoxedVertical, BoxedVertical.OnValuesChangeListener {
                 val command: UByte = MidiHelper.STATUS_PITCH_BEND or channel
                 val msg = ubyteArrayOf(0U, PITCHCENTERU)
                 MainActivity.mMidi.send(command, msg.toByteArray())
-                value = PITCHCENTER.toInt()
+                value = PITCHCENTER
             }
             else if (mRZ) {
                 val channel = if (mChannel != MidandpadDB.DEFAULT_CHANNEL) mChannel.toUByte()
